Cap02.6 CRUD com o H2 Database

+++ 15 - Spring02 - Musica/musica/src/main/java/Rios/Vinicius/musica/entities/Estado.java
@@ -10,7 +10,7 @@ public class Estado {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     @OneToMany(mappedBy = "estado")
@@ -19,16 +19,16 @@ public class Estado {
     public Estado() {
     }
 
-    public Estado(Integer id, String nome) {
+    public Estado(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
